requirements.txt
accelerate==1.8.1
aiofiles==24.1.0
albumentations==1.4.0
annotated-types==0.7.0
anyio==4.9.0
bitsandbytes==0.46.1
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
fastapi==0.116.0
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.5.1
ftfy==6.3.1
gradio==5.24.0
gradio_client==1.8.0
gradio_pdf==0.0.22
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jiwer==4.0.0
joblib==1.5.1
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
llama_cpp_python==0.3.12
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.95.1
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
orjson==3.10.18
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
psutil==5.9.8
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.8.0
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qudida==0.0.4
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.12.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spaces==0.37.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.17
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typer==0.16.0
typhoon-ocr==0.3.8
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
